+++ client/src/components/admin/Pagination/Pagination.types.ts
